bill update

+++ update.php
@@ -1,65 +1,35 @@
 <?php
-//Update
+// print_r($_POST);
+include 'config/database.php';
 
-//Deleting existing rows
 
-//
+$data = $_POST;
+
+$billNo = $data['billNo'];
+
+// die();
+
+// Deleting Data
 $dataDelete = "DELETE FROM `invoicedata` WHERE billNo = '$billNo'";
+$qryDelete = mysqli_query($conn, $dataDelete);
 
-$rowDeleted = mysqli_query($conn, $dataDelete);
-
-if ($rowDeleted) {
-    echo "Deleted";
+if ($qryDelete) {
+    $detailsDelete = "DELETE FROM `invoicedetails` WHERE billNo = '$billNo'";
+    $qryDelete2 = mysqli_query($conn, $detailsDelete);
 } else {
     echo "Error";
 }
 
-$detailsDelete = "DELETE FROM `invoicedetails` WHERE billNo = '$billNo'";
-
-$rowDeleted = mysqli_query($conn, $detailsDelete);
-
-if ($rowDeleted) {
-    echo "Deleted";
-} else {
-    echo "Error";
-}
 
 
 
 // array
 $array = $_POST;
-// print_r($array);
-echo "<br>";
-echo "<br>";
-// echo $itemName = $_POST['name1'];
-
-echo "<br>";
 
 
 
 // object
 $object = (object)$array;
-
-// print_r($object);
-
-// echo "<br>";
-// echo "<br>";
-// echo $itemName = $object->name1;
-
-// echo "<br>";
-// echo "<br>";
-
-
-
-
-$billNo = $_POST['billNo'];
-
-echo "<br>";
-echo "<br>";
-echo "<br>";
-echo "<br>";
-echo "<br>";
-echo "<br>";
 
 $numrow = $_POST['numrow'];
 
@@ -82,23 +52,11 @@ for ($i = 1; $i <= $numrow; $i++) {
 
 
     // query to insert item data
-
-
-
     $detailsData = "INSERT INTO `invoicedetails`(`billNo`,`name`,`price`,`qty`,`total`,`discount`,`disper`,`netTotal`) VALUES ('$billNo','$itemName','$itemPrice','$itemQty','$itemTotal','$itemDis','$itemDisper','$itemNetTotal')";
 
 
     $detailsInsert = mysqli_query($conn, $detailsData);
-
-    if ($detailsInsert) {
-        echo "Data Saved";
-    } else {
-        echo "Error";
-    }
 }
-
-
-
 
 
 $billTotal = $_POST['total_2'];
@@ -106,17 +64,11 @@ $billDis = $_POST['dis_2'];
 $billDisPer = $_POST['disper_2'];
 $billNetTotal = $_POST['net_total'];
 
-
-// including DB file
-
-
 $billQuery = "INSERT INTO `invoicedata`(`billNo`, `total`, `discount`, `discountPer`, `netTotal`) VALUES ('$billNo','$billTotal','$billDis','$billDisPer','$billNetTotal')";
 $billInsert = mysqli_query($conn, $billQuery);
 
 if ($billInsert) {
-    echo "Invoice Saved";
-    // header("Location: http://localhost/table/search.php");
-    // die();
+    echo "Success";
 } else {
     echo "Error";
 }
